+++ src/compiler/linker.h
@@ -388,7 +388,7 @@ class LibraryMethod {
       case frontend::FUNC_TYPE: {
         name = L'(';
         vector<frontend::Type*> func_params = type->GetFunctionParameters();
-        for(int i = 0; i < func_params.size(); ++i) {
+        for(size_t i = 0; i < func_params.size(); ++i) {
           name += EncodeUserType(func_params[i]);
         }
         name += L") ~ ";
